Move Utility/TypeConverter headers to include directory
