+++ lib/Conversion/TritonGPUToLLVM/AssertOpToLLVM.cpp
@@ -1,6 +1,6 @@
-#include "Utility.h"
 #include "mlir/Dialect/ControlFlow/IR/ControlFlowOps.h"
 #include "triton/Conversion/TritonGPUToLLVM/PatternTritonGPUOpToLLVM.h"
+#include "triton/Conversion/TritonGPUToLLVM/Utility.h"
 namespace {
 using namespace mlir;
 using namespace mlir::triton;
