+++ lib/Conversion/TritonGPUToLLVM/DotOpToLLVM/FMA.cpp
@@ -1,4 +1,4 @@
-#include "Utility.h"
+#include "triton/Conversion/TritonGPUToLLVM/Utility.h"
 
 using namespace mlir;
 using namespace mlir::triton;
