+++ lib/Conversion/TritonGPUToLLVM/ElementwiseOpToLLVM.cpp
@@ -1,8 +1,8 @@
 #include "triton/Conversion/TritonGPUToLLVM/ElementwiseOpToLLVM.h"
-#include "Utility.h"
 #include "mlir/Conversion/LLVMCommon/Pattern.h"
 #include "mlir/Dialect/GPU/IR/GPUDialect.h"
 #include "triton/Conversion/TritonGPUToLLVM/PatternTritonGPUOpToLLVM.h"
+#include "triton/Conversion/TritonGPUToLLVM/Utility.h"
 #include "triton/Dialect/TritonGPU/IR/Dialect.h"
 
 using namespace mlir::triton::gpu;
