+++ lib/Conversion/TritonGPUToLLVM/MemoryOpToLLVM.cpp
@@ -1,5 +1,5 @@
-#include "Utility.h"
 #include "triton/Conversion/TritonGPUToLLVM/PatternTritonGPUOpToLLVM.h"
+#include "triton/Conversion/TritonGPUToLLVM/Utility.h"
 
 namespace {
